two critical updates
(1) The data type of OIMS_header is obviously compound and not text, even though the data type of the the values of the different components tends to be text.
(2) the value for multiple in metadata schema should be TRUE. You can have multiple schemas . The first schema takes precedence over the second schema in case there are terms that are defined multiple times.