updating to version 2.2.0.0
The key reason for a major new sub-version is that we ran into an issue with Status. Status is most often used for condition in this case {active, inactive, depreciated, under review} whereas we were using it for requirement level. This has been changed throughout. with the attribute RequirementLevel being used for the old status and adding an optional StatusCondition attribute.